refactor: Refactor HelpText and TextLink to TS
