--- packages/forma-36-react-components/src/components/TextLink/TextLink.js
@@ -1,89 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import cn from 'classnames';
-import { iconName } from '../Icon/constants';
-import Icon from '../Icon';
-import TabFocusTrap from '../TabFocusTrap';
-import styles from './TextLink.css';
-
-class TextLink extends React.Component {
-  static propTypes = {
-    children: PropTypes.node.isRequired,
-    href: PropTypes.string,
-    disabled: PropTypes.bool,
-    testId: PropTypes.string,
-    linkType: PropTypes.oneOf([
-      'primary',
-      'positive',
-      'negative',
-      'secondary',
-      'muted',
-    ]),
-    onClick: PropTypes.func,
-    extraClassNames: PropTypes.string,
-    icon: PropTypes.oneOf(Object.keys(iconName)),
-  };
-
-  static defaultProps = {
-    linkType: 'primary',
-    href: undefined,
-    testId: 'cf-ui-text-link',
-    disabled: false,
-    onClick: undefined,
-    extraClassNames: undefined,
-    icon: undefined,
-  };
-
-  renderIcon(icon, linkType) {
-    if (!icon) return undefined;
-
-    return (
-      <div className={styles['TextLink__icon-wrapper']}>
-        <Icon
-          icon={icon}
-          color={linkType}
-          extraClassNames={styles.TextLink__icon}
-        />
-      </div>
-    );
-  }
-
-  render() {
-    const {
-      children,
-      href,
-      linkType,
-      testId,
-      onClick,
-      disabled,
-      extraClassNames,
-      icon,
-      ...otherProps
-    } = this.props;
-
-    const classNames = cn(styles.TextLink, extraClassNames, {
-      [styles[`TextLink--${linkType}`]]: linkType,
-      [styles['TextLink--disabled']]: disabled,
-    });
-
-    const Element = href ? 'a' : 'button';
-
-    return (
-      <Element
-        className={classNames}
-        data-test-id={testId}
-        onClick={!disabled ? onClick : null}
-        href={!disabled ? href : null}
-        disabled={disabled}
-        {...otherProps}
-      >
-        <TabFocusTrap>
-          {this.renderIcon(icon, linkType)}
-          {children}
-        </TabFocusTrap>
-      </Element>
-    );
-  }
-}
-
-export default TextLink;
